// usdt-test/tasks/utils/input.types.ts
export type All = {
	uri: string;
	tokenid: string;
	spender: string;
	holder: string;
	address: string;
	from: string;
	to: string;
	value: number;
	currvalue: number;
	newvalue: number;
	year: number;
	month: number;
	quantity: number;
	min: number;
	max: number;
	role: string;
	adminRole: string;
};